DRY mute filtering and combinator cleanup
