+++ packages/core/src/utils/formatQuery/defaultRuleGroupProcessorElasticSearch.ts
@@ -38,7 +38,7 @@ export const defaultRuleGroupProcessorElasticSearch: RuleGroupProcessor<Record<s
     }
 
     const processedRules = rg.rules
-      .filter(rule => typeof rule === 'string' || !rule.muted)
+      .filter(rule => !rule.muted)
       .map(rule => {
         if (isRuleGroup(rule)) {
           return processRuleGroup(rule);
