+++ packages/core/src/utils/formatQuery/defaultRuleGroupProcessorJsonLogic.ts
@@ -42,7 +42,7 @@ export const defaultRuleGroupProcessorJsonLogic: RuleGroupProcessor<RQBJsonLogic
     }
 
     const processedRules = rg.rules
-      .filter(rule => typeof rule === 'string' || !rule.muted)
+      .filter(rule => !rule.muted)
       .map(rule => {
         if (isRuleGroup(rule)) {
           return processRuleGroup(rule);
