+++ packages/core/src/utils/formatQuery/defaultRuleGroupProcessorMongoDBQuery.ts
@@ -47,7 +47,7 @@ export const defaultRuleGroupProcessorMongoDBQuery: RuleGroupProcessor = (
     let hasChildRules = false;
 
     const expressions: Record<string, unknown>[] = rg.rules
-      .filter(rule => typeof rule === 'string' || !rule.muted)
+      .filter(rule => !rule.muted)
       .map(rule => {
         if (isRuleGroup(rule)) {
           const processedRuleGroup = processRuleGroup(rule);
