+++ packages/core/src/utils/formatQuery/defaultRuleGroupProcessorSQL.ts
@@ -2,6 +2,7 @@ import type { RuleGroupProcessor, RuleGroupTypeAny } from '../../types';
 import { isRuleGroup, isRuleGroupType } from '../isRuleGroup';
 import { isRuleOrGroupValid } from '../isRuleOrGroupValid';
 import { getOption } from '../optGroupUtils';
+import { filterRulesAndCleanupCombinators } from './utils';
 
 /**
  * Default rule processor used by {@link formatQuery} for "sql" format.
@@ -32,41 +33,8 @@ export const defaultRuleGroupProcessorSQL: RuleGroupProcessor<string> = (ruleGro
       return outermostOrLonelyInGroup ? fallbackExpression : /* istanbul ignore next */ '';
     }
 
-    // First filter out muted items
-    const filteredRules = rg.rules.filter(rule => {
-      // Filter out muted rules and groups
-      if (typeof rule !== 'string' && rule.muted) {
-        return false;
-      }
-      return true;
-    });
-
-    // Clean up orphaned combinators in IC format
-    const cleanedRules = !isRuleGroupType(rg)
-      ? (() => {
-          const result: typeof filteredRules = [];
-          let lastWasRule = false;
-
-          for (let i = 0; i < filteredRules.length; i++) {
-            const current = filteredRules[i];
-
-            if (typeof current === 'string') {
-              // Only add combinator if the last item was a rule and there's a rule after this
-              const nextIsRule = i + 1 < filteredRules.length && typeof filteredRules[i + 1] !== 'string';
-              if (lastWasRule && nextIsRule) {
-                result.push(current);
-                lastWasRule = false;
-              }
-            } else {
-              // It's a rule or group
-              result.push(current);
-              lastWasRule = true;
-            }
-          }
-
-          return result;
-        })()
-      : filteredRules;
+    // Filter out muted items and clean up orphaned combinators
+    const cleanedRules = filterRulesAndCleanupCombinators(rg);
 
     const processedRules = cleanedRules
       .map(rule => {
