+++ packages/core/src/utils/formatQuery/defaultRuleGroupProcessorSequelize.ts
@@ -50,7 +50,7 @@ export const defaultRuleGroupProcessorSequelize: RuleGroupProcessor<WhereOptions
     let hasChildRules = false;
 
     const expressions: Record<string, unknown>[] = rg.rules
-      .filter(rule => typeof rule === 'string' || !rule.muted)
+      .filter(rule => !rule.muted)
       .map(rule => {
         if (isRuleGroup(rule)) {
           const processedRuleGroup = processRuleGroup(rule);
